+++ src/test/java/org/rookit/dm/album/AlbumFactoryTest.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  ******************************************************************************/
-package album;
+package org.rookit.dm.album;
 
 import static org.junit.Assert.*;
