+++ src/test/java/org/rookit/dm/artist/ArtistFactoryTest.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  ******************************************************************************/
-package artist;
+package org.rookit.dm.artist;
 
 import static org.junit.Assert.*;
